test(core,blocks): split four multi-describe files into one-describe-per-file
Per project convention "one describe per file at most." Four files
covered:

- permutations-normalize.test.ts → -gating + -dispatch (disjoint setup)
- variance-analysis.test.ts → -reference + -layered + -edge-cases (each
  owns its own beforeAll for the project it loads)
- detail-overlay.browser.test.tsx → -focus-lifecycle + -focus-trap +
  -dismissal + shared _detail-overlay-helpers.tsx
- token-navigator-keyboard.browser.test.tsx → -roving-tabindex +
  -arrow-navigation + -expand-collapse + -activation + shared
  _token-navigator-keyboard-helpers.tsx

Helpers follow the existing leading-underscore convention. No assertion
changes; same tests, same shapes, flatter reporter grouping.
